Adding in Food truck calendar events
I noticed that Chuck's provides a calendar of food trucks for each of their locations.

I thought it'd be cool to add those (a few days at a time) into the app so you'd know what was there and what was coming
over the next few days (currently set to 5 days). If there are events to show in the next 5 days, they'll load. Tapping the entry will take you to the food truck website, or if not available a helpful google search for the food truck.

From what I can tell using the Google rest APIs requires a limited list of json converters, so I went ahead and added GSON for this. I couldn't get it to really work well w/ the existing json stuff :(  There are Google API client libraries but a quick scan seemed to imply you needed to enable/use oauth, and reading JSON seemed easier.